@@ -1,13 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Version>1.4</Version>
-    <LangVersion>9</LangVersion>
+    <Version>1.4.2</Version>
+    <LangVersion>10</LangVersion>
     <Nullable>enable</Nullable>
     <Authors>d4n3436</Authors>
     <NeutralLanguage>en</NeutralLanguage>
     <TargetFrameworks>net461;netstandard2.0;netstandard2.1</TargetFrameworks>
-    <Configurations>Debug;Release;DebugLabs;ReleaseLabs</Configurations>
     <IncludeSymbols>true</IncludeSymbols>
     <SymbolPackageFormat>snupkg</SymbolPackageFormat>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
@@ -15,7 +14,6 @@
     <RepositoryUrl>https://github.com/d4n3436/Fergun.Interactive</RepositoryUrl>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageTags>discord;discordapp;discord-bot;discordbot;discord-net;interactive</PackageTags>
-    <RestoreAdditionalProjectSources>https://www.myget.org/F/discord-net/api/v3/index.json</RestoreAdditionalProjectSources>
     <Description>
 An addon that adds interactive functionality to commands.
 This is a fork of Discord.InteractivityAddon that adds several features like more customization and support for interactions.
@@ -23,30 +21,8 @@ This is a fork of Discord.InteractivityAddon that adds several features like mor
     <PackageReleaseNotes>https://github.com/d4n3436/Fergun.Interactive/releases</PackageReleaseNotes>
   </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)'=='ReleaseLabs' Or '$(Configuration)'=='DebugLabs'">
-    <PackageId>$(MSBuildProjectName).Labs</PackageId>
-    <AssemblyName>$(MSBuildProjectName).Labs</AssemblyName>
-    <DefineConstants>$(DefineConstants)TRACE;DNETLABS</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)'=='ReleaseLabs'">
-    <Optimize>True</Optimize>
-  </PropertyGroup>
-
-  <Choose>
-    <When Condition="'$(Configuration)'=='Release' Or '$(Configuration)'=='Debug'">
-      <ItemGroup>
-        <PackageReference Include="Discord.Net.WebSocket" Version="3.0.0" />
-      </ItemGroup>
-    </When>
-    <Otherwise>
-      <ItemGroup>
-        <PackageReference Include="Discord.Net.Labs.WebSocket" Version="3.6.1" />
-      </ItemGroup>
-    </Otherwise>
-  </Choose>
-
   <ItemGroup>
+    <PackageReference Include="Discord.Net.WebSocket" Version="3.4.1" />
     <PackageReference Include="Nullable" Version="1.3.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
