@@ -1,47 +1,46 @@
 using System;
 using Discord;
 
-namespace Fergun.Interactive
+namespace Fergun.Interactive;
+
+/// <summary>
+/// Specifies the actions that will be applied to a message after a timeout or a cancellation.
+/// </summary>
+[Flags]
+public enum ActionOnStop
 {
     /// <summary>
-    /// Specifies the actions that will be applied to a message after a timeout or a cancellation.
+    /// Do nothing.
     /// </summary>
-    [Flags]
-    public enum ActionOnStop
-    {
-        /// <summary>
-        /// Do nothing.
-        /// </summary>
-        None = 0,
+    None = 0,
 
-        /// <summary>
-        /// Modify the message using <see cref="IInteractiveElement{TOption}.TimeoutPage"/> or <see cref="IInteractiveElement{TOption}.CanceledPage"/>.
-        /// </summary>
-        /// <remarks>This action is mutually exclusive with <see cref="DeleteMessage"/>.</remarks>
-        ModifyMessage = 1 << 0,
+    /// <summary>
+    /// Modify the message using <see cref="IInteractiveElement{TOption}.TimeoutPage"/> or <see cref="IInteractiveElement{TOption}.CanceledPage"/>.
+    /// </summary>
+    /// <remarks>This action is mutually exclusive with <see cref="DeleteMessage"/>.</remarks>
+    ModifyMessage = 1 << 0,
 
-        /// <summary>
-        /// Delete the reactions/buttons/select menu from the message.
-        /// </summary>
-        /// <remarks>
-        /// This action is mutually exclusive with <see cref="DisableInput"/>.<br/>
-        /// If reactions are used as input, this requires the <see cref="ChannelPermission.ManageMessages"/> permission.
-        /// </remarks>
-        DeleteInput = 1 << 1,
+    /// <summary>
+    /// Delete the reactions/buttons/select menu from the message.
+    /// </summary>
+    /// <remarks>
+    /// This action is mutually exclusive with <see cref="DisableInput"/>.<br/>
+    /// If reactions are used as input, this requires the <see cref="ChannelPermission.ManageMessages"/> permission.
+    /// </remarks>
+    DeleteInput = 1 << 1,
 
-        /// <summary>
-        /// Disable the buttons or the select menu from the message. Only applicable to messages using buttons or select menus.
-        /// </summary>
-        /// <remarks>This action is mutually exclusive with <see cref="DeleteInput"/>.</remarks>
-        DisableInput = 1 << 2,
+    /// <summary>
+    /// Disable the buttons or the select menu from the message. Only applicable to messages using buttons or select menus.
+    /// </summary>
+    /// <remarks>This action is mutually exclusive with <see cref="DeleteInput"/>.</remarks>
+    DisableInput = 1 << 2,
 
-        /// <summary>
-        /// Delete the message.
-        /// </summary>
-        /// <remarks>
-        /// This action takes the highest precedence over any other flag.<br/>
-        /// Ephemeral messages can't be deleted.
-        /// </remarks>
-        DeleteMessage = 1 << 3
-    }
+    /// <summary>
+    /// Delete the message.
+    /// </summary>
+    /// <remarks>
+    /// This action takes the highest precedence over any other flag.<br/>
+    /// Ephemeral messages can't be deleted.
+    /// </remarks>
+    DeleteMessage = 1 << 3
 }
