@@ -2,104 +2,103 @@ using System;
 using System.Threading.Tasks;
 using Discord.WebSocket;
 
-namespace Fergun.Interactive
+namespace Fergun.Interactive;
+
+/// <summary>
+/// Represents an event handler with a filter.
+/// </summary>
+/// <typeparam name="TInput">The type of the incoming inputs.</typeparam>
+internal sealed class FilteredCallback<TInput> : IInteractiveCallback<TInput>
 {
-    /// <summary>
-    /// Represents an event handler with a filter.
-    /// </summary>
-    /// <typeparam name="TInput">The type of the incoming inputs.</typeparam>
-    internal sealed class FilteredCallback<TInput> : IInteractiveCallback<TInput>
+    private bool _disposed;
+
+    public FilteredCallback(Func<TInput, bool> filter, Func<TInput, bool, Task> action,
+        TimeoutTaskCompletionSource<(TInput?, InteractiveStatus)> timeoutTaskSource, DateTimeOffset startTime)
     {
-        private bool _disposed;
+        Filter = filter;
+        Action = action;
+        TimeoutTaskSource = timeoutTaskSource;
+        StartTime = startTime;
+    }
 
-        public FilteredCallback(Func<TInput, bool> filter, Func<TInput, bool, Task> action,
-            TimeoutTaskCompletionSource<(TInput?, InteractiveStatus)> timeoutTaskSource, DateTimeOffset startTime)
+    /// <summary>
+    /// Gets the filter.
+    /// </summary>
+    public Func<TInput, bool> Filter { get; }
+
+    /// <summary>
+    /// Gets the action which gets executed to incoming inputs.
+    /// </summary>
+    public Func<TInput, bool, Task> Action { get; }
+
+    /// <summary>
+    /// Gets the <see cref="TimeoutTaskCompletionSource{TResult}"/> used to set the result of the callback.
+    /// </summary>
+    public TimeoutTaskCompletionSource<(TInput?, InteractiveStatus)> TimeoutTaskSource { get; }
+
+    /// <inheritdoc/>
+    public DateTimeOffset StartTime { get; }
+
+    /// <inheritdoc/>
+    public void Cancel() => TimeoutTaskSource.TryCancel();
+
+    /// <inheritdoc/>
+    public async Task ExecuteAsync(TInput input)
+    {
+        bool success = Filter(input);
+        await Action(input, success).ConfigureAwait(false);
+
+        if (success)
         {
-            Filter = filter;
-            Action = action;
-            TimeoutTaskSource = timeoutTaskSource;
-            StartTime = startTime;
+            TimeoutTaskSource.TrySetResult((input, InteractiveStatus.Success));
+            Dispose();
+        }
+    }
+
+    /// <inheritdoc/>
+    public Task ExecuteAsync(SocketMessage message)
+    {
+        if (message is TInput input)
+        {
+            return ExecuteAsync(input);
         }
 
-        /// <summary>
-        /// Gets the filter.
-        /// </summary>
-        public Func<TInput, bool> Filter { get; }
+        throw new ArgumentException("Cannot execute this callback using a message.", nameof(message));
+    }
 
-        /// <summary>
-        /// Gets the action which gets executed to incoming inputs.
-        /// </summary>
-        public Func<TInput, bool, Task> Action { get; }
-
-        /// <summary>
-        /// Gets the <see cref="TimeoutTaskCompletionSource{TResult}"/> used to set the result of the callback.
-        /// </summary>
-        public TimeoutTaskCompletionSource<(TInput?, InteractiveStatus)> TimeoutTaskSource { get; }
-
-        /// <inheritdoc/>
-        public DateTimeOffset StartTime { get; }
-
-        /// <inheritdoc/>
-        public void Cancel() => TimeoutTaskSource.TryCancel();
-
-        /// <inheritdoc/>
-        public async Task ExecuteAsync(TInput input)
+    /// <inheritdoc/>
+    public Task ExecuteAsync(SocketReaction reaction)
+    {
+        if (reaction is TInput input)
         {
-            bool success = Filter(input);
-            await Action(input, success).ConfigureAwait(false);
-
-            if (success)
-            {
-                TimeoutTaskSource.TrySetResult((input, InteractiveStatus.Success));
-                Dispose();
-            }
+            return ExecuteAsync(input);
         }
 
-        /// <inheritdoc/>
-        public Task ExecuteAsync(SocketMessage message)
-        {
-            if (message is TInput input)
-            {
-                return ExecuteAsync(input);
-            }
+        throw new ArgumentException("Cannot execute this callback using a reaction.", nameof(reaction));
+    }
 
-            throw new ArgumentException("Cannot execute this callback using a message.", nameof(message));
+    /// <inheritdoc/>
+    public Task ExecuteAsync(SocketInteraction interaction)
+    {
+        if (interaction is TInput input)
+        {
+            return ExecuteAsync(input);
         }
 
-        /// <inheritdoc/>
-        public Task ExecuteAsync(SocketReaction reaction)
-        {
-            if (reaction is TInput input)
-            {
-                return ExecuteAsync(input);
-            }
+        throw new ArgumentException("Cannot execute this callback using an interaction.", nameof(interaction));
+    }
 
-            throw new ArgumentException("Cannot execute this callback using a reaction.", nameof(reaction));
+    /// <inheritdoc/>
+    public void Dispose() => Dispose(true);
+
+    private void Dispose(bool disposing)
+    {
+        if (_disposed) return;
+        if (disposing)
+        {
+            TimeoutTaskSource.TryDispose();
         }
 
-        /// <inheritdoc/>
-        public Task ExecuteAsync(SocketInteraction interaction)
-        {
-            if (interaction is TInput input)
-            {
-                return ExecuteAsync(input);
-            }
-
-            throw new ArgumentException("Cannot execute this callback using an interaction.", nameof(interaction));
-        }
-
-        /// <inheritdoc/>
-        public void Dispose() => Dispose(true);
-
-        private void Dispose(bool disposing)
-        {
-            if (_disposed) return;
-            if (disposing)
-            {
-                TimeoutTaskSource.TryDispose();
-            }
-
-            _disposed = true;
-        }
+        _disposed = true;
     }
 }
