@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace Fergun.Interactive
+namespace Fergun.Interactive;
+
+/// <summary>
+/// Provides an elapsed time.
+/// </summary>
+public interface IElapsed
 {
     /// <summary>
-    /// Provides an elapsed time.
+    /// Gets the elapsed time.
     /// </summary>
-    public interface IElapsed
-    {
-        /// <summary>
-        /// Gets the elapsed time.
-        /// </summary>
-        TimeSpan Elapsed { get; }
-    }
+    TimeSpan Elapsed { get; }
 }
