@@ -1,28 +1,27 @@
-namespace Fergun.Interactive
+namespace Fergun.Interactive;
+
+/// <summary>
+/// Specifies the possible status of an <see cref="IInteractiveResult{TStatus}"/> whose status is <see cref="InteractiveStatus"/>.
+/// </summary>
+public enum InteractiveStatus
 {
     /// <summary>
-    /// Specifies the possible status of an <see cref="IInteractiveResult{TStatus}"/> whose status is <see cref="InteractiveStatus"/>.
+    /// The interactive action status is unknown.
     /// </summary>
-    public enum InteractiveStatus
-    {
-        /// <summary>
-        /// The interactive action status is unknown.
-        /// </summary>
-        Unknown,
+    Unknown,
 
-        /// <summary>
-        /// The interactive action was successful.
-        /// </summary>
-        Success,
+    /// <summary>
+    /// The interactive action was successful.
+    /// </summary>
+    Success,
 
-        /// <summary>
-        /// The interactive action timed out.
-        /// </summary>
-        Timeout,
+    /// <summary>
+    /// The interactive action timed out.
+    /// </summary>
+    Timeout,
 
-        /// <summary>
-        /// The interactive action was canceled.
-        /// </summary>
-        Canceled
-    }
+    /// <summary>
+    /// The interactive action was canceled.
+    /// </summary>
+    Canceled
 }
