@@ -1,13 +1,12 @@
-namespace Fergun.Interactive.Pagination
+namespace Fergun.Interactive.Pagination;
+
+/// <summary>
+/// Represents a lazy paginator.
+/// </summary>
+public sealed class LazyPaginator : BaseLazyPaginator
 {
-    /// <summary>
-    /// Represents a lazy paginator.
-    /// </summary>
-    public sealed class LazyPaginator : BaseLazyPaginator
+    internal LazyPaginator(LazyPaginatorBuilder properties)
+        : base(properties)
     {
-        internal LazyPaginator(LazyPaginatorBuilder properties)
-            : base(properties)
-        {
-        }
     }
 }
