@@ -1,26 +1,25 @@
-namespace Fergun.Interactive.Pagination
+namespace Fergun.Interactive.Pagination;
+
+/// <summary>
+/// Represents a builder class for making a <see cref="LazyPaginator"/>.
+/// </summary>
+public sealed class LazyPaginatorBuilder : BaseLazyPaginatorBuilder<LazyPaginator, LazyPaginatorBuilder>
 {
     /// <summary>
-    /// Represents a builder class for making a <see cref="LazyPaginator"/>.
+    /// Initializes a new instance of the <see cref="LazyPaginatorBuilder"/> class.
     /// </summary>
-    public sealed class LazyPaginatorBuilder : BaseLazyPaginatorBuilder<LazyPaginator, LazyPaginatorBuilder>
+    public LazyPaginatorBuilder()
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="LazyPaginatorBuilder"/> class.
-        /// </summary>
-        public LazyPaginatorBuilder()
+    }
+
+    /// <inheritdoc/>
+    public override LazyPaginator Build()
+    {
+        if (Options.Count == 0)
         {
+            WithDefaultEmotes();
         }
 
-        /// <inheritdoc/>
-        public override LazyPaginator Build()
-        {
-            if (Options.Count == 0)
-            {
-                WithDefaultEmotes();
-            }
-
-            return new LazyPaginator(this);
-        }
+        return new LazyPaginator(this);
     }
 }
