@@ -1,33 +1,32 @@
-namespace Fergun.Interactive.Pagination
+namespace Fergun.Interactive.Pagination;
+
+/// <summary>
+/// Specifies the paginator actions related to emotes.
+/// </summary>
+public enum PaginatorAction
 {
     /// <summary>
-    /// Specifies the paginator actions related to emotes.
+    /// Go to the next page.
     /// </summary>
-    public enum PaginatorAction
-    {
-        /// <summary>
-        /// Go to the next page.
-        /// </summary>
-        Forward,
+    Forward,
 
-        /// <summary>
-        /// Go to the previous page.
-        /// </summary>
-        Backward,
+    /// <summary>
+    /// Go to the previous page.
+    /// </summary>
+    Backward,
 
-        /// <summary>
-        /// Skip to the end.
-        /// </summary>
-        SkipToEnd,
+    /// <summary>
+    /// Skip to the end.
+    /// </summary>
+    SkipToEnd,
 
-        /// <summary>
-        /// Skip to the start.
-        /// </summary>
-        SkipToStart,
+    /// <summary>
+    /// Skip to the start.
+    /// </summary>
+    SkipToStart,
 
-        /// <summary>
-        /// Exit the paginator.
-        /// </summary>
-        Exit
-    }
+    /// <summary>
+    /// Exit the paginator.
+    /// </summary>
+    Exit
 }
