@@ -1,15 +1,14 @@
 using System.Collections.Generic;
 
-namespace Fergun.Interactive.Pagination
+namespace Fergun.Interactive.Pagination;
+
+/// <summary>
+/// Represents the properties of a <see cref="BaseStaticPaginatorBuilder{TPaginator, TBuilder}"/>.
+/// </summary>
+public interface IBaseStaticPaginatorBuilderProperties : IBasePaginatorBuilderProperties
 {
     /// <summary>
-    /// Represents the properties of a <see cref="BaseStaticPaginatorBuilder{TPaginator, TBuilder}"/>.
+    /// Gets or sets the pages of the paginator.
     /// </summary>
-    public interface IBaseStaticPaginatorBuilderProperties : IBasePaginatorBuilderProperties
-    {
-        /// <summary>
-        /// Gets or sets the pages of the paginator.
-        /// </summary>
-        IList<IPageBuilder> Pages { get; set; }
-    }
+    IList<IPageBuilder> Pages { get; set; }
 }
