@@ -1,26 +1,25 @@
-namespace Fergun.Interactive.Pagination
+namespace Fergun.Interactive.Pagination;
+
+/// <summary>
+/// Represents a builder class for making a <see cref="StaticPaginator"/>.
+/// </summary>
+public sealed class StaticPaginatorBuilder : BaseStaticPaginatorBuilder<StaticPaginator, StaticPaginatorBuilder>
 {
     /// <summary>
-    /// Represents a builder class for making a <see cref="StaticPaginator"/>.
+    /// Initializes a new instance of the <see cref="StaticPaginatorBuilder"/> class.
     /// </summary>
-    public sealed class StaticPaginatorBuilder : BaseStaticPaginatorBuilder<StaticPaginator, StaticPaginatorBuilder>
+    public StaticPaginatorBuilder()
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="StaticPaginatorBuilder"/> class.
-        /// </summary>
-        public StaticPaginatorBuilder()
+    }
+
+    /// <inheritdoc/>
+    public override StaticPaginator Build()
+    {
+        if (Options.Count == 0)
         {
+            WithDefaultEmotes();
         }
 
-        /// <inheritdoc/>
-        public override StaticPaginator Build()
-        {
-            if (Options.Count == 0)
-            {
-                WithDefaultEmotes();
-            }
-
-            return new StaticPaginator(this);
-        }
+        return new StaticPaginator(this);
     }
 }
