@@ -1,14 +1,13 @@
-namespace Fergun.Interactive.Selection
+namespace Fergun.Interactive.Selection;
+
+/// <summary>
+/// Represents a selection of options.
+/// </summary>
+/// <typeparam name="TOption">The type of the options.</typeparam>
+public class Selection<TOption> : BaseSelection<TOption>
 {
-    /// <summary>
-    /// Represents a selection of options.
-    /// </summary>
-    /// <typeparam name="TOption">The type of the options.</typeparam>
-    public class Selection<TOption> : BaseSelection<TOption>
+    internal Selection(IBaseSelectionBuilderProperties<TOption> properties)
+        : base(properties)
     {
-        internal Selection(IBaseSelectionBuilderProperties<TOption> properties)
-            : base(properties)
-        {
-        }
     }
 }
